Unify container fixtures
